- fixed the ProbeForWrite prototype - some fixes to get ntoskrnl to build with -Wcast-qual and -Wsign-compare - don't touch any memory outside the given range in ProbeForWrite and touch all pages within the range
svn path=/trunk/; revision=19748

+++ reactos/ntoskrnl/io/pnpmgr.c
@@ -221,7 +221,7 @@ IoGetDeviceProperty(
 
         /* Query the device caps */
         Status = IopQueryDeviceCapabilities(DeviceNode, &DeviceCaps);
-        if (NT_SUCCESS(Status) && (DeviceCaps.Address != -1))
+        if (NT_SUCCESS(Status) && (DeviceCaps.Address != (ULONG)-1))
         {
             /* Return length */
             *ResultLength = sizeof(ULONG);
